Replace cluster specific envoy filter for reversed vpn with a static one.
As described in #4381, the previous approach with cluster specific envoy filters causes connections to be reset
when a new cluster is created or an existing one is deleted, i.e. in case a reconfiguration happens on the istio
ingress gateway/envoy proxy. In this change the per cluster approach is replaced with a single static envoy filter
so that no reconfigurations happen on cluster creation/deletion.